long to float fix DONE

+++ kafka-stream/src/main/scala/org/esgi/project/streaming/models/Likes.scala
@@ -3,7 +3,7 @@ import play.api.libs.json.{Json, OFormat}
 
 case class Likes(
                   _id: Int,
-                  score: Long
+                  score: Float
                 )
 
 object Likes{
